Updated Camera check
Updated the camera to fix the bug where the player could be lost sight of if they backed up during a room transition. However to apply this fox to all rooms the camera zone code will need to be prefabed and the enemies will also have to be adjusted to use the new system.