Full Camera rework
Updated all of the camera zones to use the new system, removing the camera de-sync bug. Also made a few small changes like more visible floor on one of the corridors, moving the credits slightly, and making the fade in panel not block the first room while in the editor.